refactor gameStage
