+++ app/src/main/java/com/ligrim/tower_defense/EnemyType.java
@@ -1,0 +1,9 @@
+package com.ligrim.tower_defense;
+
+public enum EnemyType {
+    NONE,
+    TANKER_ENEMY,
+    NORMAL_ENEMY,
+    SMALLER_ENEMY,
+    BOSS_ENEMY;
+}
