+++ app/src/main/java/com/ligrim/tower_defense/base/BFS.java
@@ -45,13 +45,11 @@ public class BFS {
     public static List<Position> getRouteFromMatrix(int[][] mapDat) {
         reset();
         queue = new LinkedList<>();
-        Point start = new Point(0, 0);
 
         for (int i = 0; i < mapDat.length; ++i) {
             for (int j = 0; j < mapDat[0].length; ++j) {
                 if (mapDat[i][j] == START) {
                     queue.add(new Point(j, i));
-                    start = new Point(j, i);
                     break;
                 }
             }
@@ -59,9 +57,9 @@ public class BFS {
         bfs(mapDat);
         // search for the nearest target
         Point nearest = findNearestTarget(mapDat);
-        /*System.out.println(nearest.x + " " + nearest.y);
+        System.out.println(nearest.x + " " + nearest.y);
         print(mapDat);
-        print(map);*/
+        print(map);
 
         LinkedList<Position> result = new LinkedList<>();
         Stack<Point> stack = new Stack<>();
@@ -144,6 +142,8 @@ public class BFS {
     }
 
     private static void smoothCurve(LinkedList<Position> list) {
+
+        Position previousDeletedNode = list.get(0);
         for (int i = 0; i + 2 < list.size(); ++i) {
             // find right angle
             if (Math.abs(Position.distanceSquared(list.get(i), list.get(i + 1)) +
@@ -158,6 +158,7 @@ public class BFS {
                 if (Position.distance(firstMid, list.get(i + 1)) < GameStage.UNIT_HEIGHT * .325f ) {
                     firstMid = list.get(i);
                     patch = false;
+                    origin = Position.midPoint(previousDeletedNode, list.get(i + 2));
                 }
 
                 Position firstVec = new Position(firstMid.getX() - origin.getX(), firstMid.getY() - origin.getY());
@@ -171,6 +172,7 @@ public class BFS {
                 float unit_angle = diffAngle / curveDegree;
                 /*System.out.println("diffangle: " + diffAngle / Math.PI * 180);*/
 
+                previousDeletedNode = list.get(i + 1);
                 list.remove(i++ + 1);
                 for (int j = 0; j <= curveDegree; ++j) {
                     if (!patch && j == 0) continue;
@@ -202,9 +204,9 @@ public class BFS {
                 map[i][j] = sc.nextInt();
             }
         }
-        //print(map);
-        //getRouteFromMatrix(map);
-        //print(BFS.map);
+        //print(Map);
+        //getRouteFromMatrix(Map);
+        //print(BFS.Map);
         List<Position> p = getRouteFromMatrix(map);
         for (Position i : p) {
             System.out.print(i.getX() + " " + i.getY() + ", ");
